requirements.txt
streamlit==1.30.0
langchain==0.1.0
langchain-community==0.0.10
langchain-nvidia-ai-endpoints==0.0.8
PyPDF2==3.0.1
chromadb==0.4.20
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
Pillow>=11.2.1
pytesseract>=0.3.13
pdf2image>=1.17.0
